Fixed BRK issue with tune 2 of Monty on the Run
At the end of tune 2 the player jumped to execute invalid code, which
included a 0x00 BRK instruction.  The player immediately stopped with an
error and didn't continue to play the remaining buffered data.  It's unknown
whether this is a problem with the SID file or whether it relies on extra
game code on the real machine.

On a real C64 (and PC sidplay) the BRK jumps to the kernal handler, which
jumps somewhere else invalid and remains stuck in a loop.  Playback has
already finished so everything appears normal.

The work-around changes BRK to be treated as RET, so it returns from the
playback routine.  The interrupt handler plays the rest of the buffered data
and everything seems fine.